+++ app/stores/eawsRegions.ts
@@ -1,6 +1,7 @@
 import outline_properties from "@eaws/outline_properties/index.json";
 import { LanguageSchema } from "../appStore";
 import { z } from "zod/mini";
+import { LatLngBounds } from "leaflet";
 
 export const RegionOutlineSchema = z.object({
   id: z.string(),
@@ -45,4 +46,24 @@ export function eawsRegionIds(): string[] {
   return eawsRegions
     .map(properties => properties.id)
     .filter(id => !config.regionsRegex.test(id));
+}
+
+export function eawsRegionsBounds(
+  regionCodes: string[],
+  f: (
+    regionCode: string
+  ) => undefined | { bbox: [number, number, number, number] } = eawsRegion
+): LatLngBounds {
+  return regionCodes.reduce((b, r) => {
+    const region = f(r);
+    if (region?.bbox) {
+      b.extend(
+        new LatLngBounds([
+          [region.bbox[1], region.bbox[0]],
+          [region.bbox[3], region.bbox[2]]
+        ])
+      );
+    }
+    return b;
+  }, new LatLngBounds([]));
 }