+++ app/views/blogPost.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from "react";
 import "@albina-euregio/linea";
+import "@albina-euregio/linea/aws-stats";
 import { useIntl } from "../i18n";
 import PageHeadline from "../components/organisms/page-headline";
 import HTMLPageLoadingScreen from "../components/organisms/html-page-loading-screen";
-import SmShare from "../components/organisms/sm-share";
 import HTMLHeader from "../components/organisms/html-header";
 import TagList from "../components/blog/tag-list";
 import { DATE_TIME_FORMAT } from "../util/date";
@@ -122,7 +122,6 @@ const BlogPost = () => {
         <section className="panel blog-post">{content}</section>
       </section>
       {renderLinkToBlogOverview()}
-      <SmShare />
     </>
   );
 };