+++ app/views/weather.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { useIntl } from "../i18n";
 import PageHeadline from "../components/organisms/page-headline";
-import SmShare from "../components/organisms/sm-share";
 import HTMLHeader from "../components/organisms/html-header";
 import WeatherMap from "../components/weather/weather-map";
 import FeatureInfo from "../components/weather/feature-info";
@@ -51,7 +50,7 @@ const Weather = () => {
     if (currentTime && domainId && !params.timestamp) {
       redirectPage($router, "weatherMapDomainTimestamp", {
         domain: domainId,
-        timestamp: currentTime.toISOString(),
+        timestamp: currentTime.toString(),
         timeSpan: store.timeSpan.get()
       });
     }
@@ -87,7 +86,6 @@ const Weather = () => {
           </div>
         )}
       </section>
-      <SmShare />
     </>
   );
 };